nav bar done

+++ backend/controllers/product.controller.js
@@ -41,13 +41,17 @@ export const createProduct = async(req, res) => {
 export const deleteProductExtra = async(req, res) => {
     const { id } = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Product not found" });
+    }
+
     try {
 
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted successfully" });
 
     } catch (error) {
-        return res.status(404).json({ success: false, message: "Product not found" });
+        return res.status(500).json({ success: false, message: " Internal server error " });
     }
 };
